Use hand-written FromJSON instance
numerous problems have turned up because we were relying on an
automatic FromJSON instance. The first is that a refactor to this type
caused code in production not to parse, the second is that the
unscoped label "kjDir" led to shadowing issues in var names.

Generated instances work great when you don't care about terseness,
but for a publicly exposed config file, this is a better solution.